+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Report diagnostic instead of panicking for mismatched return types
 - Report diagnostic for function argument number length mismatches
 - Report diagnostic for attempting to call non-function values
+- [Report `const` declarations with values not known at compile time](https://github.com/amp-lang/ampc/issues/9)
 
 ### [`snapshot-02-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-02-06-2023-a)
 (initial snapshot)